Reworked Documentation.
