+++ python/README.md
@@ -1,60 +1,58 @@
-<p align="center">
-<img src="https://www.intrepidcs.com/wp-content/uploads/2018/03/IntrepidCS_logo.png" width="350"/>
+# neoradio2 Python library
+![Intrepid Control Systems, Inc.](IntrepidCS_logo.png)
 
-</p>
+# Documentation
 
-
-# RAD-IO2 Python API
-
-The RAD-IO2 series is a family of ruggedized products that provide an isolated analog, digital or temperature interface to a PC via the PC’s USB port. These tools can also be paired with Intrepid products that include a USB port such as neoVI ION, neoVI FIRE 2, RAD-Galaxy, and RAD-Gigalog. In addition, the RAD-IO2-CANHUB can power and convert the native UART signal to CAN or CAN FD for use in any CAN device.
-
-The RAD-IO2 family communicates on an open source UART based serial communication protocol. Up to eight devices can be daisy chained. The chain length is limited by current supplied to the chain through USB. All RAD-IO2 devices have input to output isolation, and 2.5kV isolation between each of the eight banks. Bank to bank isolation is important because it allows the common mode voltage of each input signal to be different than the other channels in other banks. (This is a major source of measurement error and can damage to the product.) Additionally, noise on one channel will not affect other channels.
-
-The list of RAD-IO2 devices supported by the API are as follows:
-
-* [RAD-IO2-TC](https://www.intrepidcs.com/products/analog-daq-devices/rad-io2-series/)
-* [RAD-IO2-AIN](https://www.intrepidcs.com/products/analog-daq-devices/rad-io2-series/)
-* [RAD-IO2-AOUT](https://www.intrepidcs.com/products/analog-daq-devices/rad-io2-series/)
-* [RAD-IO2-PWRRLY](https://www.intrepidcs.com/products/analog-daq-devices/rad-io2-series/)
-* [RAD-IO2-DIO](https://www.intrepidcs.com/products/analog-daq-devices/rad-io2-series/)
-* [RAD-IO2-CANHUB](https://www.intrepidcs.com/products/analog-daq-devices/rad-io2-series/)
-* [RAD-IO2-BADGE](https://www.intrepidcs.com/products/analog-daq-devices/rad-io2-series/)
+https://libneoradio2.readthedocs.io/
 
 
+# Quick Start
 
-### Prerequisities
-<a href="https://www.python.org/downloads/">Python 3.5 or Higher running on your local system </a> <br>
+This is intended for more advanced users with intention of quickly starting off. For more details on installation please visit the documentation at https://libneoradio2.readthedocs.io/.
 
-## Installation
-Install from PyPi using [pip](http://www.pip-installer.org/en/latest/), a
-package manager for Python.
+```pip install neoradio2```
+
+**Note:** Windows `pip.exe` is usually located under the Scripts directory under the Python installation directory.
+
+
+# Quick Example
+```
+#!/usr/bin/env python3
+import neoradio2
+
+if __name__ == "__main__":
+    # Open and close all the attached neoRAD-IO2 devices
+    for device in neoradio2.find():
+        print(device)
+        handle = neoradio2.open(device)
+        # Do stuff here
+        neoradio2.close(handle)
 
 ```
-pip install neoradio2
-```
-Don't have pip installed? Try installing it, by running this from the command
-line:
-```
-$ curl https://raw.github.com/pypa/pip/master/contrib/get-pip.py | python
-```
 
-You may need to run the above commands with `sudo`.
+# neoRAD-IO2 Product Line
+
+The neoRAD-IO2 series is a family of ruggedized products that provide an isolated analog, digital or temperature interface to a PC via the PC’s USB port. These tools can also be paired with Intrepid products that include a USB port such as neoVI ION, neoVI FIRE 2, RAD-Galaxy, and RAD-Gigalog. In addition, the neoRAD-IO2-CANHUB can power and convert the native UART signal to CAN or CAN FD for use in any CAN device.
+
+The neoRAD-IO2 family communicates on an open source UART based serial communication protocol. Up to eight devices can be daisy chained. The chain length is limited by current supplied to the chain through USB. All neoRAD-IO2 devices have input to output isolation, and 2.5kV isolation between each of the eight banks. Bank to bank isolation is important because it allows the common mode voltage of each input signal to be different than the other channels in other banks. (This is a major source of measurement error and can damage to the product.) Additionally, noise on one channel will not affect other channels. For more details please visit https://www.intrepidcs.com/products/analog-daq-devices/rad-io2-series/
 
 
-## Getting Started
+# neoRAD-IO2-BADGE Quick Start
 
-* RAD-IO2-BADGE: Demo Board For Multichannel Analog Measurement
-The RAD-IO2-Badge is a demonstration platform that can measure multichannel analog input and digital output on a single low-cost device. The device mimics the RAD-IO2-AIN and RAD-IO2-PWRRLY.
+neoRAD-IO2-BADGE is a demo board For multi-channel analog measurement tool. The device mimics the neoRAD-IO2-AIN and neoRAD-IO2-PWRRLY Products.
 
-* A Tour of RAD-IO2-BADGE Hardware - please [click here](docs/badge/readme/Badge_HW_GUIDE.md)
-* For guide on getting started with Analog Inputs using the python API for RAD-IO2-BADGE - please [click here](docs/badge/readme/Badge_AIN_GUIDE.md)
-* For guide on Digital Output using the python API for RAD-IO2-BADGE  - please [click here](docs/badge/readme/Badge_DO_GUIDE.md)
+* Examples can be found [here](example/badge/) 
 
+## License - MIT
 
+<img src="https://opensource.org/files/OSI_Approved_License.png" width="185" height="254">
+Copyright <2019> <Intrepid Control Systems, Inc.>
 
-## License
+Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
 
-This project is licensed under the BSD License - see the [LICENSE.md](LICENSE.md) file for details
+The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
