+++ python/build_all.bat
@@ -1,0 +1,36 @@
+@ECHO OFF
+mode con cols=140 lines=70
+
+rmdir /s /q build
+
+mkdir old_dist
+move dist\* old_dist\
+
+REM C:\Python27\scripts\pip install wheel twine pybind11 --upgrade
+REM C:\Python33\scripts\pip install wheel twine pybind11 --upgrade
+REM C:\Python34\scripts\pip install wheel twine pybind11 --upgrade
+REM C:\Python35\scripts\pip install wheel twine pybind11 --upgrade
+C:\Python36-32\scripts\pip install wheel twine pybind11 --upgrade
+C:\Python37-32\scripts\pip install wheel twine pybind11 --upgrade
+
+REM C:\Python27-64\scripts\pip install wheel twine pybind11 --upgrade
+REM C:\Python34-64\scripts\pip install wheel twine pybind11 --upgrade
+REM C:\Python35-64\scripts\pip install wheel twine pybind11 --upgrade
+C:\Python36-64\scripts\pip install wheel twine pybind11 --upgrade
+C:\Python37-64\scripts\pip install wheel twine pybind11 --upgrade
+
+REM C:\Python27\python setup.py sdist bdist_wheel
+REM C:\Python33\python setup.py sdist bdist_wheel
+REM C:\Python34\python setup.py sdist bdist_wheel
+REM C:\Python35\python setup.py sdist bdist_wheel
+C:\Python36-32\python setup.py sdist bdist_wheel
+C:\Python37-32\python setup.py sdist bdist_wheel
+
+
+REM C:\Python27-64\python setup.py sdist bdist_wheel
+REM C:\Python34-64\python setup.py sdist bdist_wheel
+REM C:\Python35-64\python setup.py sdist bdist_wheel
+C:\Python36-64\python setup.py sdist bdist_wheel
+C:\Python37-64\python setup.py sdist bdist_wheel
+
+REM C:\Python36-32\scripts\twine upload dist/* -r pypitest
